requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.14.0
pillow>=9.5.0
nltk>=3.8.1
tqdm>=4.66.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.16.0
peft>=0.7.0
bitsandbytes>=0.41.0
deepspeed>=0.12.0
flash-attn>=2.3.0
pycocotools>=2.0.7